refactor: consolidate Profile/Settings/Privacy into tabbed /settings page
Merges three pages into one /settings route with four tabs:
- Profile: name, bio, contact, availability, location, skills
- Account: change email, change password, delete account
- Notifications: email digest preference
- Privacy & Data: consent settings, data export, link to /privacy policy

/profile now redirects to /settings. Header consolidates "My Profile" +
"Account Settings" into a single "Settings" link. /privacy retains the
full policy text; delete account card removed (now in Account tab).
Local group accepted notification link updated to /settings.
