+++ app/settings/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next'
 
-export const metadata: Metadata = { title: 'Account Settings' }
+export const metadata: Metadata = { title: 'Settings' }
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return <>{children}</>
