+++ components/Header.tsx
@@ -242,16 +242,10 @@ export default function Header() {
                       onClick={() => setUserMenuOpen(false)}
                     >
                       <Link
-                        href="/profile"
-                        className="block px-4 py-3 text-brand-text no-underline"
-                      >
-                        My Profile
-                      </Link>
-                      <Link
                         href="/settings"
                         className="block px-4 py-3 text-brand-text no-underline"
                       >
-                        Account Settings
+                        Settings
                       </Link>
                       <Link
                         href="/privacy"
@@ -426,8 +420,7 @@ export default function Header() {
                       </span>
                     )}
                   </MobileNavLink>
-                  <MobileNavLink href="/profile">My Profile</MobileNavLink>
-                  <MobileNavLink href="/settings">Account Settings</MobileNavLink>
+                  <MobileNavLink href="/settings">Settings</MobileNavLink>
 
                   {user.isAdmin && (
                     <>
